requirements.txt
langchain
pandas
langchain-google-genai
python-dotenv
streamlit